完善 增删改查 最后剩余 Home逻辑

+++ src/components/profile/config/ChangePassword.vue
@@ -46,36 +46,19 @@ export default {
       this.$refs[formName].validate((valid) => {
         if (valid) {
           this.isLoading = true
-          post('resetPassword', {
-            'password': this.ruleForm.oldPassWord,
-            'resetPassword': this.ruleForm.newPassWord,
-          }, {
-            headers: {
-              'Content-Type': 'application/json',
-              'Accept': 'application/json',
-              'Authorization': 'Bearer ' + localStorage.access_token
-            }
-          }).then(response => {
-            this.$message({message: response.data.message, type: 'success'})
-            localStorage.clear()
-            this.$router.push({name: 'Login',})
-            this.isLoading = false
-          }).catch(error => {
-            if (error.status === 403) {
-              this.$message({message: error.data.message, type: 'error'})
-            }
-            if (error.status === 422) {
-              this.$message({message: Object.values(error.data.errors)[0][0], type: 'error'})
-            }
-            if (error.status === 500) {
-              this.$message({message: '服务器错误，请重试。', type: 'error'})
-            }
-            //超时处理
-            if (error === 'timeout') {
-              this.$message({message: '请求超时，请重试，或检查网络。', type: 'error'});
-            }
-            this.isLoading = false
-          })
+          post('resetPassword',
+              {'password': this.ruleForm.oldPassWord, 'resetPassword': this.ruleForm.newPassWord},
+              {headers: {'Content-Type': 'application/json', 'Accept': 'application/json', 'Authorization': 'Bearer ' + localStorage.access_token}})
+              .then(response => {
+                this.isLoading = false
+                if (response.status === 200) {
+                  localStorage.clear()
+                  this.$router.push({name: 'Login',})
+                }
+              })
+              .catch(error => {
+                this.isLoading = false
+              })
         } else {
           return false;
         }
